Generalize target-environment-variables.
It will be useful to stuff other kind of environment variables inside,
so change its meaning a little.

* antioxidant-packages.scm (target-environment-variables): Rename to
...
(rust-environment-variables): ... this, don't consider RUSTC_BOOTSTRAP
special but separate it a bit from the rest.
(antioxidant-build): Adjust to new variable names.
* antioxidant.scm (set-platform-dependent-variables): Rename to ...
(set-rust-environment-variables): ... this, and adjust docstring.
(%standard-antioxidant-phases): Adjust to new variable names.